requirements.txt
Flask==1.1.2
awscli==1.18.23
boto3==1.12.23
boto==2.49.0
pycryptodome==3.9.7
mysql-connector-python==8.0.22
